+++ housekeeping/src/main/java/com/hotels/housekeeping/model/EntityLegacyReplicaPath.java
@@ -46,6 +46,12 @@ public abstract class EntityLegacyReplicaPath implements LegacyReplicaPath {
   @Column(name = "path_event_id", nullable = true, length = 250)
   protected String pathEventId;
 
+  @Column(name = "metastore_database_name")
+  protected String metastoreDatabaseName;
+
+  @Column(name = "metastore_table_name")
+  protected String metastoreTableName;
+
   protected EntityLegacyReplicaPath() {}
 
   @Override
@@ -93,6 +99,26 @@ public abstract class EntityLegacyReplicaPath implements LegacyReplicaPath {
     this.creationTimestamp = creationTimestamp;
   }
 
+  @Override
+  public String getMetastoreDatabaseName() {
+    return metastoreDatabaseName;
+  }
+
+  @Override
+  public void setMetastoreDatabaseName(String metastoreDatabaseName) {
+    this.metastoreDatabaseName = metastoreDatabaseName;
+  }
+
+  @Override
+  public String getMetastoreTableName() {
+    return metastoreTableName;
+  }
+
+  @Override
+  public void setMetastoreTableName(String metastoreTableName) {
+    this.metastoreTableName = metastoreTableName;
+  }
+
   @PrePersist
   protected void onPersist() {
     setCreationTimestamp(new Instant().getMillis());
@@ -107,6 +133,8 @@ public abstract class EntityLegacyReplicaPath implements LegacyReplicaPath {
         .add("pathEventId", pathEventId)
         .add("path", path)
         .add("creationTimestamp", creationTimestamp)
+        .add("metastoreDatabaseName", metastoreDatabaseName)
+        .add("metastoreTableName", metastoreTableName)
         .toString();
   }